FTP deployment revision
Added browserify
Flattened folder structure in dist
Added FTP deployment
removed heroku deployment
