+++ .gitignore
@@ -1,2 +1,4 @@
 *.log
 *.swp
+dist
+node_modules
