+++ .travis.yml
@@ -1,6 +1,11 @@
 language: node_js
 node_js:
 - '4.1'
+env:
+  global:
+  - secure: aCBBTcKIKFP4RfitBYyAfr72xzE2NzzRY2WEvbHtrsnf1K2EFebOTuxtfesZI8wUV3we56JcAufq7z7DTwMyCcTsTPxJKONNzg73f9cNdwnBN7n7UWS8TjfCVvCkN5c1MPOFjsqgWw1N0bY9t3H/rNb4ewQmlW4EIMNWaQVvZSJYtUVX8TFdwlxrsldyihx/eWIFkZ0g/Q2ljXcdyFqXbSSGQlYjkTH6Xsgxg781EWDvwJPaXUR+u78rdXZ4QvHjyydYF5Ty+r4qeBLqYUSG1kykVooXx4p998XsDGg5H5N+huh6cTxjixfEN6UahKbhoVQ8upc94LTxYfMDbXAi7BSe3ARisJeMASwJrYLLOSZTcGN01ZDu9LcYYi1KmVDQEQlDjJMn8U6yvyt9SrV9L1+k0q3TmW+bXAwa9HYcZLLPxLLnoIVi8TBXP8VODlXDUqxaaTnVVX0teJrtyqEqFVDGUR7IfDyug6+U0fblF6ElTttehBFSOhaC9yFTkZUVRTpqeNEw+M6O6B7umSIDCTu1SJs4S0zSocT7aOTq+XQE/RhHbv2t4mUt+bZLybcs8ftEbrhcjOczwkF7KG8xCbyHdg3tTBOP5Zzs/bdRmaYZk3PAiwWagHjImgv1RQ2eTrX1O9H2fMxjA5BYqwapskAdGosSilKlauYtI90L9rg=
+  - secure: AvIPZZ07u8hm5z81q3nwm9sTH959bbEgq0y+u39er/Co/fSOca/BmdyjV9Dmq4l07OythPM2S2hLDswbY78/NYHmesbc/a1vaqL2CELlQvHC73awEn3dybRhudIaaEXwsf3kMgNOdhdvEb4HX3Oy86/vxePJ8N34TvfDqQRvaIDTad/JonO2sKWqd6ZzwzbKRsHf7Wc7POxV/xKVoIEPcQjeXU0/OnlA0O3TU/UKHXyyqWbERRYo24Tg+i9AyCxGoDVBZSC1uglwBjlgi7unOKSdvJ+Uo2NnhYMiphZ0NPN8YkNbmOKsbvQfVgUep+hOddACrYGPuBXKaAauvEqk88vc3Tdvln5O+pX4vEk0wbVUcY7hR4XFIpjWBU9LZzjoGAMf4DTicJWHjS1iDKkPthDIPC1LH6hNmPsr9qo7OoRrdwwutnifAsMoWJ6MzyLDZ1KjvSs2gJQ8wyOAPd3WGabWlhO/iQygHqlMOI+UgOTA0as4TYUzqLLlFygib3eCad2w89NCZl+hx9t1QhEWGNSlMvupdpAcujPDP3On3ZhqASKeu2arDtz4Q9Djlg+7mo5LobHwQRqq1hguBMFfFYtxu6wAGy+uCE4ZYDbrrPnXGaaL+UtMv9DDehJTxPxFkdiG8jNBCwrkdiFRcJknW0lGCGbt6lCHHB8bJEnJGFs=
+  - secure: A0yAY0QVvVxknYjpy+VXTmR0DUOVgL92aScW+XD6FuhR6yqv3IhmGpaDgEuo1IFBiyeBZ9E32GFXyFMWrnziA1Qc6hU7hM0zq8oFCL15bNyc5a9yzccDVRT2Z4cKLoCNi/Sl+21g7wGhBA8ggsoA6OGD7mQWFlZZ1yzC+fViUSE3SFdOY0rm40BbycewbFG0N03WRXm0yOMI3sQpSYtm5URqSAUR3HHuY7bVb05RuABzN3lZFZUpS1gRdZ1Vz39zwx6/C7OORgJSWH1nQAXavWzamhlNp52K89HT7uEjpsoQoTx9dealZMLVKIZM8GbljnBfFMEXmzDKsf+TEodn7kcwQa6zQyY8j3xoKOQjpE8T6/2qjEZMy4Nnc6FKVJ8H8K/acVDJ1ztb63NZx2D6ZZUCGgpvg3Zsp1kkuxehu3jwAHa4ZHKgRdCFCTkpu2jG3cYnfb64diBIV7G6JqF/HN90ORZGyEYG1Md+w/nG/56gjlNVH3uD/zqQXrNgeloCwvvSjzOCDNY7o1ixq1HbREXHX5jIynQTaK1Dj1pm17CjJBSqwMdnslDKh4XZTf9ktICCpNsGD8DD7CPMpaAFdz8gvU5GCkzCepV/5j/mLse0TITfq+/U5FMH1s1vAFJheDzATRkVJ0+RbGsbkTDmgJxZoH8lsJJBWRn2h+GQkDc=
 before_script:
 - gulp build
 before_deploy:
@@ -8,11 +13,8 @@ before_deploy:
 - rm -rf src
 - rm -rf test
 deploy:
-  provider: heroku
-  api_key:
-    secure: ECnQqn/L1MzL7ePifqEVbflcfoCyYvgQ374P/YnsTQx6Pj1mpuQD8Ty/f5yxe2fQQrXhJWyQ99KKBvpmjIzjaV2Z3ye5W5diHEe9u8d94EL65a0pAWkZ9j/a4jVNVBbo+VMJjrD14FmI/3iuVosa9/S0F2byB0SBl6wYSLtHtNP3DuerfOnkKkw2++Ngx2LtiydxAggFVnN6kERc/avmh8i55y/xgbMMjdvzqwq+4+zRm/1vydMoSOzzmtjl8Jc5k8CMndzOoiAMXtPl0qTTzR4whlVO5vp4cHsSfOtmZhp4BuwbkbkTvsiG9Y5tz+UBNDgl9DD4aF+k5VqppGL+iiSakZRrUKB3HWNp7YAnILufzdXVZld4xNpb7mMoBo7WpVq1aMLc2kpho023UqG0O/d+RBNGUXLnweAvVYgeXf4p4Bzu3XtPcwvgFo0ev/ZlEa5xutOhou3lXznRr31TmEIzkUS/Iy8SlvH8FHWH5o2VShK9f/bC2p+xWxgKCd/D/yA/GeTYehGjDQoDBYQ+rA5+5GfB4sgl7F7VAbpWvw7qRStC59KUAtaNAbVhWx4H64sp3VwA6rURp7oTGnyVZB+x612jTENAF6z1Yrjk8dPBNxP0AUFjPYYQFVbo629zm4xUuagSgzKflr0y35sYiIYBss2CC2I93j5oFk1ELb8=
-  app: 
-    develop: stormy-caverns-1711
+- provider: script
+  script: scripts/deploy.sh
   on:
-    repo: Bangsadrengur/squealing-octo-rutabaga
+    branch: master
   skip_cleanup: true
