+++ gulpfile.js
@@ -2,6 +2,9 @@ var gulp = require('gulp');
 var html5Lint = require('gulp-html5-lint');
 var del = require('del');
 var eslint = require('gulp-eslint');
+var browserify = require('browserify');
+var source = require('vinyl-source-stream');
+var gutil = require('gulp-util');
 
 gulp.task('js-lint', function () {
   return gulp.src(['src/js/*.js'])
@@ -16,15 +19,21 @@ gulp.task('js-lint', function () {
     .pipe(eslint.failAfterError());
 });
 
-// Disabled until gulp-html5-lint is fixed
+gulp.task('js', function() {
+  return browserify('./src/js/app.js').
+    bundle().
+    on('error', function (e) {}).
+    pipe(source('bundle.js')).
+    pipe(gulp.dest('./dist/'));
+});
+
 gulp.task('html5-lint', function() {
-  return;
-  //return gulp.src('./src/*.html')
-  //  .pipe(html5Lint());
+  return gulp.src('./src/*.html')
+    .pipe(html5Lint());
 });
 
 gulp.task('watch-src', function() {
-  gulp.watch('src/js/*.js', ['js-lint', 'js-to-dist']);
+  gulp.watch('src/js/*.js', ['js-lint', 'js']);
   gulp.watch('src/css/*.css', ['css-to-dist']);
   gulp.watch('src/css/*.woff', ['fonts-to-dist']);
   gulp.watch('src/*.html', ['html5-lint', 'html-to-dist']);
@@ -32,62 +41,29 @@ gulp.task('watch-src', function() {
   gulp.watch('src/assets/*', ['assets-to-dist']);
 });
 
-gulp.task('mithril-to-dist', function() {
-  var mithrilPath = './node_modules/mithril';
-  gulp.src([
-    mithrilPath + '/mithril.min.js',
-    mithrilPath + '/mithril.min.js.map',
-    mithrilPath + '/mithril.js'
-    ])
-    .pipe(gulp.dest('./dist/lib/mithril'));
-});
-
-gulp.task('bootstrap-to-dist', function() {
-  var bootstrapPath = './node_modules/bootstrap/dist/**/*';
-  gulp.src(bootstrapPath).pipe(
-    gulp.dest('./dist/lib/bootstrap')
-  );
-});
-
-gulp.task('js-to-dist', function() {
-  gulp.src('./src/js/*.js')
-    .pipe(gulp.dest('./dist/js'));
-});
-
 gulp.task('html-to-dist', function() {
   gulp.src('./src/**/*.html')
     .pipe(gulp.dest('./dist'));
 });
 
-gulp.task('jquery-to-dist', function() {
-  gulp.src('./node_modules/jquery/dist/*.js').pipe(
-    gulp.dest('./dist/lib/jquery')
-  );
-});
-
-gulp.task('jquery-cycle-to-dist', function() {
-  gulp.src('./src/jquery/*.js')
-    .pipe(gulp.dest('./dist/jquery'));
-});
-
 gulp.task('css-to-dist', function() {
   gulp.src('./src/css/*.css')
-    .pipe(gulp.dest('./dist/css'));
+    .pipe(gulp.dest('./dist'));
 });
 
 gulp.task('fonts-to-dist', function() {
   gulp.src('./src/css/*.woff')
-    .pipe(gulp.dest('./dist/css'));
+    .pipe(gulp.dest('./dist'));
 });
 
 gulp.task('images-to-dist', function() {
   gulp.src('./src/images/*')
-    .pipe(gulp.dest('./dist/images'));
+    .pipe(gulp.dest('./dist'));
 });
 
 gulp.task('assets-to-dist', function() {
   gulp.src('./src/assets/*')
-    .pipe(gulp.dest('./dist/assets'));
+    .pipe(gulp.dest('./dist'));
 });
 
 gulp.task('clean', function() {
@@ -95,16 +71,12 @@ gulp.task('clean', function() {
 });
 
 gulp.task('build', [
-  'js-to-dist',
+  'js',
   'html-to-dist',
-  'jquery-to-dist',
-  'jquery-cycle-to-dist',
   'css-to-dist',
   'fonts-to-dist',
   'images-to-dist',
-  'assets-to-dist',
-  'mithril-to-dist',
-  'bootstrap-to-dist'
+  'assets-to-dist'
 ]);
 
 gulp.task('watch', ['watch-src']);
