+++ package.json
@@ -28,17 +28,21 @@
     "express": "^4.13.3"
   },
   "devDependencies": {
-    "bootstrap": "^3.3.5",
+    "browserify": "^13.0.1",
     "del": "^2.0.2",
     "gulp": "^3.9.0",
     "gulp-eslint": "^1.0.0",
     "gulp-html5-lint": "^1.0.1",
+    "gulp-util": "^3.0.7",
     "jasmine-core": "^2.3.4",
     "jquery": "^1.9.1",
     "karma": "^0.13.14",
+    "karma-browserify": "^5.0.5",
     "karma-jasmine": "^0.3.6",
     "karma-phantomjs-launcher": "^0.2.1",
     "mithril": "^0.2.0",
-    "phantomjs": "^1.9.18"
+    "phantomjs": "^2.1.7",
+    "vinyl-source-stream": "^1.1.0",
+    "watchify": "^3.7.0"
   }
 }
