+++ src/index.html
@@ -1,17 +1,13 @@
 <!DOCTYPE HTML>
 <head>
   <meta charset="UTF-8">
-  <script src="lib/mithril/mithril.min.js"></script>
-  <script src="lib/jquery/jquery.min.js"></script>
-  <script src="jquery/jquery.cycle2.js"></script>
-  <script src="jquery/jquery.cycle2.center.min.js"></script>
-  <link href="css/stylesheet.css" rel="stylesheet" type="text/css"/>
+  <link href="stylesheet.css" rel="stylesheet" type="text/css"/>
   <title>Suzukipíanóskólinn</title>
 </head>
 <body>
   <div id="header">
     <a href="#">
-      <img id="logo" src="images/logo.png" alt="Logo"/>
+      <img id="logo" src="logo.png" alt="Logo"/>
     </a>
     <div id="header-slideshow"></div>
   </div>
@@ -22,6 +18,5 @@
     <p>Suzukipíanóskólinn - Langarima 21, 112 Reykjavík - Sími 8637963 - 
     www.suzukipianoskolinn.is - suzukipianoskolinn@gmail.com</p>
   </div>
-  <script src="js/app.js"></script>
-  <script src="js/router.js"></script>
+  <script src="bundle.js"></script>
 </body>
