--- test/dummySpec.js
@@ -1,5 +1,0 @@
-describe('first test', function() {
-  it('should succeed', function() {
-    expect(true).toBeTruthy();
-  });
-});
